Fix Darkmode and introduce Themecontext
